Integrate v8.29 fromFRK

+++ repo/1.clone.ps1
@@ -11,12 +11,20 @@ param (
 . $fbv # NOTE: should do that for all private functions
 
 $verNew = Find-BlitzVersion -LatestFromGitHub
-$verActual = Find-BlitzVersion -Path $PSScriptRoot\frk\sp_Blitz.sql
-Write-Host ("New version = {0}, Actual = {1}" -f $verNew.Version, $verActual.Version)
+if (Test-Path $PSScriptRoot\frk\sp_Blitz.sql) {
+    $verActual = Find-BlitzVersion -Path $PSScriptRoot\frk\sp_Blitz.sql
+    Write-Host ("New version = {0}, Actual = {1}" -f $verNew.Version, $verActual.Version)
+} else {
+    Write-Host ("New version = {0}, could not find old one" -f $verNew.Version)
+}
+
 
 if ($verNew.Version -gt $verActual.Version) {
-    Remove-Item $PSScriptRoot\frk.old -Recurse -Force -ErrorAction SilentlyContinue
-    Rename-Item $PSScriptRoot\frk frk.old             -ErrorAction SilentlyContinue
+    if (Test-Path $PSScriptRoot\frk.old) {
+        Remove-Item $PSScriptRoot\frk.old -Recurse -Force -ErrorAction SilentlyContinue
+        Rename-Item $PSScriptRoot\frk frk.old             -ErrorAction SilentlyContinue
+    }
+
     git clone 'https://github.com/BrentOzarULTD/SQL-Server-First-Responder-Kit.git' $PSScriptRoot\frk
     Push-Location $PSScriptRoot\frk # NOTE: should do that without changing location
     git switch main
